- add authorization logic to Infusions endpoints to prevent other users from updating/deleting other users' data - also update page to pass user Id to endpoints

+++ src/pages/api/infusions/[id].ts
@@ -13,12 +13,14 @@ export default async function handler(
     return res.status(401).send("unauthorized");
   }
 
+  const userIDFromSession: number | null = session?.user?.id || null;
+
   //  handle routes
   switch (req.method) {
     case "DELETE":
-      return handleDelete(req, res);
+      return handleDelete(req, res, userIDFromSession);
     case "PUT":
-      return handleUpdate(req, res);
+      return handleUpdate(req, res, userIDFromSession);
     default:
       return res.status(405).json({ message: "Method not allowed" });
   }
@@ -28,9 +30,22 @@ const handleError = (res: NextApiResponse, e: any) => {
   res.status(500).send(e ? e?.details : "error unknown");
 };
 
-const handleDelete = async (req: NextApiRequest, res: NextApiResponse) => {
+const handleDelete = async (
+  req: NextApiRequest,
+  res: NextApiResponse,
+  userIDFromSession: number | null
+) => {
   try {
     const { id } = req.query;
+    const { userID } = req.body;
+
+    if (userIDFromSession !== Number(userID)) {
+      return res
+        .status(400)
+        .send(
+          "Error: You are not authorized to track another user's infusion."
+        );
+    }
 
     const { data: deletedInfusion, error: errorForDeletedInfusion } =
       await supabase.from("infusion").delete().match({ id: id });
@@ -44,9 +59,22 @@ const handleDelete = async (req: NextApiRequest, res: NextApiResponse) => {
   }
 };
 
-const handleUpdate = async (req: NextApiRequest, res: NextApiResponse) => {
+const handleUpdate = async (
+  req: NextApiRequest,
+  res: NextApiResponse,
+  userIDFromSession: number | null
+) => {
   try {
     const { id } = req.query;
+    const { userID } = req.body;
+
+    if (userIDFromSession !== Number(userID)) {
+      return res
+        .status(400)
+        .send(
+          "Error: You are not authorized to track another user's infusion."
+        );
+    }
 
     const { bleed_location, infusion_date, notes, treated_within } = req.body;
